# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: StudyApp API
  version: "1.0.0"
  description: Scan-to-Quiz educational platform API
servers:
  - url: http://localhost:8080
    description: Development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - UNAUTHORIZED
            - FORBIDDEN
            - NOT_FOUND
            - VALIDATION_ERROR
            - CONFLICT
            - RATE_LIMIT_EXCEEDED
            - FEATURE_NOT_AVAILABLE
            - SCAN_LIMIT_EXCEEDED
            - INTERNAL_ERROR
            - PAYMENT_FAILED
        message:
          type: string
        details:
          nullable: true
        trace_id:
          type: string
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
    MeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        email_verified:
          type: boolean
...
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password, new_password_confirm]
      properties:
        current_password:
          type: string
        new_password:
          type: string
        new_password_confirm:
          type: string
...
    FlagWithStats:
      type: object
      properties:
        key:
          type: string
        enabled:
          type: boolean
        rollout_percentage:
          type: integer
        override_count:
          type: integer
        ai_call_count_24h:
          type: integer
    AdminSetFlagRequest:
      type: object
      required: [key, enabled]
      properties:
        key:
          type: string
        enabled:
          type: boolean
        rollout_percentage:
          type: integer
...
    SelectGoalRequest:
      type: object
      required: [goalId]
      properties:
        goalId:
          type: integer
        targetDate:
          type: string
          format: date
    MyGoal:
      type: object
      properties:
        studentGoalId:
          type: integer
        goalId:
          type: integer
        name:
          type: string
        examName:
          type: string
        targetDate:
          type: string
        status:
          type: string
        progress:
          type: object
...
    SubmitPracticeRequest:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
        submittedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /auth/change-password:
    post:
      summary: Change password (authenticated)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/me:
    get:
      summary: Get current user profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MeResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /goals/select:
    post:
      summary: Select or switch career goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectGoalRequest"
      responses:
        "200":
          description: Goal selected
        "403":
          description: Feature not available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /goals/my:
    get:
      summary: Get active career goal
      responses:
        "200":
          description: Active goal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyGoal"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Abandon active career goal
      responses:
        "200":
          description: Goal abandoned
...
  /goals/my/practice/{setId}/submit:
    post:
      summary: Submit daily practice answers
      parameters:
        - name: setId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPracticeRequest"
      responses:
        "200":
          description: Practice scored
...
  /admin/features:
    get:
      summary: List feature flags with stats (admin)
      responses:
        "200":
          description: Feature flags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FlagWithStats"
    post:
      summary: Update a feature flag (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSetFlagRequest"
      responses:
        "200":
          description: Flag updated
...